backend/requirements.txt
# FastAPI + servidor
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Postgres
psycopg2-binary==2.9.10

# Parse XML rápido (extração de eventos)
lxml==5.3.0

# Env e utilidades
python-dotenv==1.0.1
